requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-extensions==2.2.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
httpie==1.0.3
idna==2.8
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
names
django-cors-headers
urllib3==1.25.3
